Make sure the format specifier matches natural data value without casting

+++ code/qcommon/cvar.c
@@ -1227,9 +1227,9 @@ void Cvar_Update(vmCvar_t* vmCvar)
     }
     vmCvar->modificationCount = cv->modificationCount;
     if (strlen(cv->string) + 1 > MAX_CVAR_VALUE_STRING)
-        Com_Error(ERR_DROP, "Cvar_Update: src %s length %u exceeds MAX_CVAR_VALUE_STRING",
+        Com_Error(ERR_DROP, "Cvar_Update: src %s length %lu exceeds MAX_CVAR_VALUE_STRING",
                   cv->string,
-                  (unsigned int)strlen(cv->string));
+                  strlen(cv->string));
     Q_strncpyz(vmCvar->string, cv->string, MAX_CVAR_VALUE_STRING);
 
     vmCvar->value = cv->value;
